🎉 feat(opencommit.yml): add OpenCommit GitHub Action (#87)
* 🎉 feat(opencommit.yml): add OpenCommit GitHub Action
The OpenCommit GitHub Action has been added to the repository. This action is triggered on push events, excluding the main, master, dev, development, and release branches. The action sets up a Node.js environment, checks out the repository, and runs the OpenCommit action using the di-sukharev/opencommit@github-action-v1.0.4 action. The OpenCommit action uses the GPT-3.5-turbo model to generate commit messages based on the changes made in the commit. The action is configured to use the GitHub token and OpenAI API key stored in the repository secrets.

* 📦 chore(opencommit.yml): downgrade opencommit action to v1.0.2 to fix compatibility issues with other dependencies

---------

Co-authored-by: juanjo96dev <juanjo96developer@gmail.com>