📝 docs(README.md): add recommendation to use fastypest in a setupFile… (#88)
* 📝 docs(README.md): improve readability and add recommendation to use fastypest in a setupFilesAfterEnv file when using Jest

* 📝 docs(README.md): add note and formatting to example of use with jest
The commit adds a note to the example of use with jest in the README.md file. The note recommends using the example in a setupFilesAfterEnv file. Additionally, the commit adds formatting to the note to make it stand out.

* 📝 docs(README.md): improve readability by adding a note and line breaks

* 🗑️ chore(opencommit.yml): remove OpenCommit workflow
The OpenCommit workflow has been removed from the repository. It was previously used to automatically generate commit messages using OpenAI's GPT-3 language model. However, it was decided that manually writing commit messages is more efficient and allows for more control over the content of the messages.

---------

Co-authored-by: juanjo96dev <juanjo96developer@gmail.com>